requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests==2.32.3
markdown==3.7
python-docx==1.1.2
pytest==8.3.3
